chore(renderer): code cleanup
also, remove all 'glide' files.

Fixes #107
